+++ epstein_files/documents/messenger_log.py
@@ -1,3 +1,4 @@
+import logging
 import re
 from collections import defaultdict
 from dataclasses import dataclass, field
@@ -15,7 +16,7 @@ from epstein_files.util.data import iso_timestamp, listify, sort_dict
 from epstein_files.util.doc_cfg import Metadata, TextCfg
 from epstein_files.util.highlighted_group import get_style_for_name
 from epstein_files.util.logging import logger
-from epstein_files.util.rich import build_table
+from epstein_files.util.rich import build_table, highlighter
 
 CONFIRMED_MSG = 'Found confirmed counterparty'
 GUESSED_MSG = 'This is probably a conversation with'
@@ -27,7 +28,12 @@ REDACTED_AUTHOR_REGEX = re.compile(r"^([-+•_1MENO.=F]+|[4Ide])$")
 class MessengerLog(Communication):
     """Class representing one iMessage log file (one conversation between Epstein and some counterparty)."""
     config: TextCfg | None = None
-    _messages: list[TextMessage] = field(default_factory=list)
+    messages: list[TextMessage] = field(default_factory=list)
+    phone_number: str | None = None
+
+    def __post_init__(self):
+        super().__post_init__()
+        self.messages = [self._build_message(match) for match in MSG_REGEX.finditer(self.text)]
 
     def first_message_at(self, name: str | None) -> datetime:
         return self.messages_by(name)[0].timestamp()
@@ -37,26 +43,28 @@ class MessengerLog(Communication):
             return None
 
         info_msg = GUESSED_MSG if self.is_attribution_uncertain() else CONFIRMED_MSG
-        author_txt = Text(self.author_or_unknown(), style=self.author_style + ' bold')
-        return Text(f"({info_msg} ", style='dim').append(author_txt).append(')')
+        author_txt = Text(self.author, style=self.author_style + ' bold')
+        txt = Text(f"({info_msg} ", style='dim').append(author_txt)
+
+        if self.phone_number:
+            txt.append(f" using the phone number {self.phone_number}")
+
+        return highlighter(txt.append(')'))
 
     def last_message_at(self, name: str | None) -> datetime:
         return self.messages_by(name)[-1].timestamp()
 
-    def messages(self) -> list[TextMessage]:
-        """Lazily evaluated accessor for self._messages."""
-        if not self._messages:
-            self._messages = [self._build_message(match) for match in MSG_REGEX.finditer(self.text)]
-
-        return self._messages
-
     def messages_by(self, name: str | None) -> list[TextMessage]:
         """Return all messages by 'name'."""
-        return [m for m in self.messages() if m.author == name]
+        return [m for m in self.messages if m.author == name]
 
     def metadata(self) -> Metadata:
         metadata = super().metadata()
-        metadata.update({'num_messages': len(self.messages())})
+        metadata.update({'num_messages': len(self.messages)})
+
+        if self.phone_number:
+            metadata['phone_number'] = self.phone_number
+
         return metadata
 
     def _border_style(self) -> str:
@@ -65,11 +73,16 @@ class MessengerLog(Communication):
     def _build_message(self, match: re.Match) -> TextMessage:
         """Turn a regex match into a TextMessage."""
         author_str = REDACTED_AUTHOR_REGEX.sub('', match.group(1).strip())
+        is_phone_number = author_str.startswith('+')
 
-        # If the Sender: is redacted that means it's from self.author
+        if is_phone_number:
+            logger.warning(f"{self.summary()} Found phone number: {author_str}")
+            self.phone_number = author_str
+
+        # If the Sender: is redacted or if it's an unredacted phone number that means it's from self.author
         return TextMessage(
-            author=self.author if (author_str.startswith('+') or not author_str) else author_str,
-            author_str=author_str if author_str.startswith('+') else None,  # Preserve phone numbers
+            author=self.author if (is_phone_number or not author_str) else author_str,
+            author_str=author_str if is_phone_number else None,  # Preserve phone numbers
             id_confirmed=not self.is_attribution_uncertain(),
             text=match.group(4).strip(),
             timestamp_str=match.group(2).strip(),
@@ -90,7 +103,7 @@ class MessengerLog(Communication):
         yield self.file_info_panel()
         yield Text('')
 
-        for message in self.messages():
+        for message in self.messages:
             yield message
 
     @classmethod
@@ -99,7 +112,7 @@ class MessengerLog(Communication):
         sender_counts: dict[str | None, int] = defaultdict(int)
 
         for message_log in imessage_logs:
-            for message in message_log.messages():
+            for message in message_log.messages:
                 sender_counts[message.author] += 1
 
         return sender_counts