+++ scripts/dump_email_timestamps.py
@@ -1,11 +1,14 @@
 #!/usr/bin/env python
 # Print email ID + timestamp
+import logging
+import sys
 from collections import defaultdict
 
 from rich.markup import escape
 from rich.panel import Panel
 
-from scripts.use_pickled import epstein_files
+from scripts.use_pickled import console, epstein_files
+from epstein_files.util.constants import ALL_FILE_CONFIGS
 from epstein_files.util.data import sort_dict
 from epstein_files.util.rich import console, print_json
 
@@ -17,6 +20,15 @@ counts = defaultdict(int)
 #     max_file_sizes[doc.class_name()] = max(max_file_sizes[doc.class_name()], doc.file_size())
 #     console.print(doc.summary())
 #     print_json('metadata', doc.metadata())
+
+for email in epstein_files.emails:
+    if email.is_local_extract_file():
+        email.warn(f'is local extract, file_id={email.file_id}, filename={email.filename}, url_slug={email.url_slug}')
+        email.warn(f"config={email.config}")
+
+
+sys.exit()
+
 
 for email in sorted(epstein_files.emails, key=lambda e: -len(e.actual_text)):
     if email.is_fwded_article() or email.is_junk_mail() or email.is_duplicate():